% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toSequence.R
\name{getActivations}
\alias{getActivations}
\title{Obtain activations of all nodes}
\usage{
getActivations(network, sequence)
}
\arguments{
\item{network}{A network}

\item{sequence}{A sequence}
}
\description{
Retrieves activations of all nodes for a given \code{Sequence}.
}
