% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getOutputs.R
\name{getOutputs}
\alias{getOutputs}
\title{Get activations of all output nodes for a given \code{sequence}}
\usage{
getOutputs(network, x, index = NULL, ...)
}
\arguments{
\item{network}{A network}

\item{x}{A \code{Sequence}}
}
\description{
Get activations of all output nodes for a given \code{sequence}
}
