% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getTargets.R
\name{getTargets}
\alias{getTargets}
\title{get targets of a given sequence}
\usage{
getTargets(x, ...)
}
\arguments{
\item{x}{A \code{Sequence}.}
}
\value{
A matrix of targets
}
\description{
get targets of a given sequence
}
