% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/banana.R
\name{\%>\%}
\alias{\%>\%}
\title{pipe command to add elements to a banana object}
\usage{
x \%>\% y
}
\description{
Typical usage is:
}
\details{
network <- Network() %>% FeedforwardEnsemble(TANH, 1) %>% FeedforwardEnsemble(TANH, 10) %>% 
FeedforwardEnsemble(TANH, 10) %>% BackpropTrainer() %>% SquaredErrorFunction()
}
