% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/banana.R
\name{has.sequenceset}
\alias{has.sequenceset}
\title{test whether a banana object has an associated sequenceset}
\usage{
has.sequenceset(x)
}
\description{
test whether a banana object has an associated sequenceset
}
