% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotActivations.R
\name{plotActivations}
\alias{plotActivations}
\title{Plot the activations across nodes for a given sequence}
\usage{
plotActivations(net, seq, node.names = NULL)
}
\description{
Plot the activations across nodes for a given sequence
}
