% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/banana.R
\name{predict.bnn}
\alias{predict.bnn}
\title{predict}
\usage{
\method{predict}{bnn}(x, newdata = NULL, index = NULL, ...)
}
\arguments{
\item{x}{A bnn object}
}
\description{
Implementation of generic \code{predict} function
}
