% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toSequence.R
\name{toSequence}
\alias{toSequence}
\title{Conversion to Sequence format}
\usage{
toSequence(x, input, target, ...)
}
\arguments{
\item{x}{a \code{"data.frame"}}

\item{input}{column id(s) of input parameters}

\item{target}{column id(s) of target parameters}
}
\value{
Sequence
}
\description{
Converts a data.frame to a \code{Sequence}
}
