% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/toSequence.R
\name{toSequenceSet}
\alias{toSequenceSet}
\title{Conversion to SequenceSet format}
\usage{
toSequenceSet(x, input, target)
}
\arguments{
\item{x}{a \code{"data.frame"}}

\item{input}{column id(s) of input parameters}

\item{target}{column id(s) of target parameters}
}
\value{
SequenceSet
}
\description{
Converts a data.frame to a \code{SequenceSet}.
}
